Move async-map into simengine
